test function of swapping

+++ polygon/sushiswap/autocompound.py
@@ -1,12 +1,13 @@
 
-from src.helper.format import convertFromWei, convertFromWeiUSDC
+from enum import auto
+from src.helper.format import convertFromWei, convertFromWeiUSDC, convertToWei, convertToWeiUSDC
 from src.constants.constants import AURUM_TOKEN_CONTRACT, RAIDER_TOKEN_CONTRACT, USDC_TOKEN_CONTRACT, WMATIC_TOKEN_CONTRACT
 from sushiswap.depositLiquidity import depositToken
 from sushiswap.swapToken import simulateAmount, swapTxn
 from src.clients import aurumStakingClient, sushiswapRouterClient
 from raider.getAurumLpRewards import claimAurumLpRewards
 from src.logger.txLogger import logTx, txLogger
-
+from raider.stakeAurumLp import stakeAurumLp
 
 raiderWmaticAurumPath = [RAIDER_TOKEN_CONTRACT,
                          WMATIC_TOKEN_CONTRACT, AURUM_TOKEN_CONTRACT]
@@ -41,11 +42,11 @@ def stakeAndSellRaider() -> None:
 def autoCompound() -> None:
     raiderAmt = claimAurumLpRewards()
     txLogger.info(f"Claimed raider amt of {raiderAmt}")
-    splitAmt = raiderAmt / 2
-
+    splitAmt = int(raiderAmt / 2)
     # swap raider to aurum
     minAmtA = swapTxn(splitAmt, raiderWmaticAurumPath)
-
     # swap raider to usdc
     minAmtB = swapTxn(splitAmt, raiderWmaticUsdcPath)
     depositToken(AURUM_TOKEN_CONTRACT, USDC_TOKEN_CONTRACT, minAmtA, minAmtB)
+    # get balance of SLP 
+    # stake SLP
